updated markup

+++ js/app/controllers/payeesCtrl.js
@@ -4,6 +4,9 @@ bankControllers.controller('PayeesCtrl', ['$scope', 'Storage', 'UI', function ($
 	};
 
 	$scope.removePayee = function(payee){
+		if (!window.confirm("Are you sure wnat to delete payee?"))
+			return;
+
 		Storage.removePayee(payee);
 		UI.showMessage("Payee removed", UI.types.success);
 	};
